tools/ml-conversion/pyproject.toml
[project]
name = "voice-cloning-test"
version = "1.0.0"
description = "Voice cloning test suite supporting Chatterbox TTS and XTTS-v2"
readme = "README.md"
requires-python = ">=3.10,<3.13"
license = { text = "MIT" }
authors = [
    { name = "Meeshy Team" }
]

dependencies = [
    # PyTorch (will be installed via TTS engines)
    "torch>=2.0.0",
    "torchaudio>=2.0.0",

    # Chatterbox TTS (default engine) - install from GitHub for multilingual
    "chatterbox-tts @ git+https://github.com/resemble-ai/chatterbox.git",

    # Audio Processing
    "soundfile>=0.12.1",
    "sounddevice>=0.4.6",
    "librosa>=0.10.0",
    "audioread>=3.0.0",
    "scipy>=1.10.0",
    "pydub>=0.25.1",

    # Speech Recognition
    "openai-whisper>=20231117",

    # Translation
    "deep-translator>=1.11.0",

    # Analysis
    "numpy>=1.24.0",
    "scikit-learn>=1.3.0",

    # Utilities
    "tqdm>=4.65.0",
    "requests>=2.28.0",
    "ffmpeg-python>=0.2.0",
]

[project.optional-dependencies]
# XTTS-v2 support (requires Python <3.12)
xtts = [
    "TTS>=0.22.0",
]

# Speaker diarization (requires HF_TOKEN)
diarization = [
    "pyannote.audio>=3.1.0",
]

# Noise reduction
denoise = [
    "noisereduce>=2.0.0",
]

# All optional features
all = [
    "TTS>=0.22.0",
    "pyannote.audio>=3.1.0",
    "noisereduce>=2.0.0",
]

[project.scripts]
voice-clone = "voice_cloning_test:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
# uv-specific configuration
dev-dependencies = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
]

[tool.hatch.build.targets.wheel]
packages = ["."]
